Moved PrimaryBalance, TokensCounter, PrimaryToken to components folder
