+++ src/components/PrimaryToken/PrimaryTokenStyles.js
@@ -1,0 +1,20 @@
+/**
+ * Copyright 2017–2018, LaborX PTY
+ * Licensed under the AGPL Version 3 license.
+ */
+
+import { StyleSheet } from 'react-native'
+
+export default StyleSheet.create({
+  balanceText: {
+    fontSize: 22,
+    fontWeight: '700',
+  },
+  balanceContainer: {
+    flexDirection: 'row',
+    marginTop: 24,
+  },
+  balanceNumber: {
+    marginLeft: 4,
+  },
+})
