+++ src/components/PrimaryToken/index.js
@@ -1,0 +1,6 @@
+/**
+ * Copyright 2017–2018, LaborX PTY
+ * Licensed under the AGPL Version 3 license.
+ */
+
+ export default from './PrimaryToken'
