+++ src/components/TokensCounter/TokenCounterStyles.js
@@ -1,0 +1,15 @@
+/**
+ * Copyright 2017–2018, LaborX PTY
+ * Licensed under the AGPL Version 3 license.
+ */
+
+import { StyleSheet } from 'react-native'
+import colors from '../../common/colors'
+
+export default StyleSheet.create({
+  tokens: {
+    color: colors.foregroundLighter,
+    fontWeight: '200',
+    marginTop: 4,
+  },
+})
