+++ src/components/TokensCounter/testData.js
@@ -1,0 +1,3 @@
+export default {
+  tokensLength: 2,
+}
